Create CanSum
Given is a targetSum and an array of numbers. It should return True if the numbers added up to give the TargetSum. Before Memoization, time complexity is O(n^m) and space is O(m), where m is the target num and n is the number of the elements in an array. After Memoization, it is actually O(n*m) time